pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "engineering-metrics-dashboard"
version = "1.0.0"
description = "Automated DORA metrics dashboard using Render Workflows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "michael@render.com" }
]
keywords = ["dora-metrics", "render-workflows", "engineering-metrics"]

dependencies = [
    "render-sdk>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
run-workflow = "metrics_dashboard.workflow:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
